requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==37.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.23.3
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.4.4
passlib==1.7.4
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
xlrd==2.0.1
xlwt==1.3.0
